Switch nvim-ufo from using treesitter to LSP
The treesitter provider didn't work well for C programs where the return
type of a function was on its own line. The function name would be
hidden in this case when collapsing all folds. Now it works like vscode.
If there is no LSP available, it will fold based on indent. This does
mean that goto labels and switch case labels don't work to well though.