Replace mini.statusline with lualine plugin
The main issue this fixes is file paths not always showing as relative.
The %f item used in the 'statusline' option is not enough to guarantee a
relative path if an absolute path was given when using the :e command to
open a file. Using %{expand('%:~:.')} fixes this issue, assuming a
normal file has been opened, but doesn't display correctly for things
like the quickfix list and unnamed buffers. The lualine plugin has its
own logic which mimics what %f does, except it actually displays
relative paths correctly (internally it uses the same expand command
given above). Do note that the Vim implementation of %f (search for
'case STL_FILEPATH' in the [Neo]Vim source code) has more complete logic
that also accounts for things like the q: (command line) window and
displays special names for those types of windows.