Add custom treesitter source to dropbar plugin
The builtin treesitter source isn't very good, so use another plugin to
provide a better treesitter source. The chosen plugin is nvim-gps. This
plugin is deprecated, but a fork exists which fixes some issues stopping
it from working. Note that if the LSP source wasn't wanted as well, then
dropbar could be replaced by the fgheng/winbar.nvim plugin and a custom
source wouldn't be needed.