Fix get_jk_mapping function for macros
The custom vertical movement is disabled when recording a macro, which
is nice, but when the macro is replayed this behavior is not disabled,
which is no good. Unfortunately I cannot find a way to detect if a macro
is currently being replayed, so just remove this behavior for now. I
guess if you're recording a macro you should always use 1j and 1k
instead of j and k if some lines could wrap or disable line wrap...