Fix up diagnostic config
Some upstream diagnostic code is not the way I want it (and some of it
is unnecessary anyway).

Remove the config I don't want.

This change does mean that now diagnostic virtual text is disabled
(since that's the new nvim default). It could be re-enabled later if I
miss it.