Force usage of tmux clipboard when inside tmux
In tmux panes that were created locally, copying text over SSH remotely
from nvim does not work. This is because of the environment variables
set for the terminal which influence which clipboard provider nvim
decides to use.

Fix this issue by forcing tmux as the clipboard provider when inside a
tmux pane.